project/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.14

# AI and ML
openai==1.86.0
google-generativeai==0.8.3
google-cloud-texttospeech==2.17.0
google-genai==1.38.0
# Environment and Config
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic-core==2.33.1

# HTTP Client
httpx==0.28.1
requests==2.32.3

# ASGI Server
uvicorn==0.34.0

# Image processing
pillow==11.0.0

python-multipart==0.0.20